+++ src/birdwatcher/video.py
@@ -144,7 +144,7 @@ class VideoFileStream():
                             ffprobepath=ffprobepath)
 
     def extract_audio(self, outputpath=None, overwrite=False, 
-                      codec='pcm_s24le', channel=None, ffmpegpath='ffmpeg', 
+                      codec='copy', channel=None, ffmpegpath='ffmpeg',
                       loglevel='quiet'):
         """Extract audio as wav file.
 
@@ -157,7 +157,9 @@ class VideoFileStream():
         overwrite : bool, default=False
             Overwrite if audio file exists or not.
         codec : str, default='pcm_s24le'
-            ffmpeg audio codec, with 24-bit pcm as default output.
+            ffmpeg audio codec, with as default copying codec to output. Another
+            choice would be 'pcm_s24le', which is a high-quality setting, but may
+            change the audio data as saved in video.
         channel : int, default=None
             Channel number to extract. The default None will extract all 
             channels.